(#47) 🔄 refactor: 메시지 말풍선 컴포넌트명 수정
