+++ src/pages/Chatbot/components/RookieMessageWithTail.tsx
